refactor: move api functions into api.go
