+++ pkg/lbaas/frontend/integration_test.go
@@ -6,6 +6,7 @@ package frontend
 import (
 	"context"
 
+	"go.anx.io/go-anxcloud/pkg/genericresource"
 	lbaasBackend "go.anx.io/go-anxcloud/pkg/lbaas/backend"
 	"go.anx.io/go-anxcloud/pkg/lbaas/common"
 
@@ -20,7 +21,7 @@ const loadbalancerIdentifier = "fc5d7390e9e4400a9efc73b4d8e0613a"
 
 var _ = Describe("lbaas/frontend client", Label("old client", "slow"), func() {
 	var cli client.Client
-	var api API
+	var api genericresource.API[Frontend, Definition]
 	var backend lbaasBackend.Backend
 
 	BeforeEach(func() {
