+++ pkg/lbaas/server/integration_test.go
@@ -7,6 +7,7 @@ import (
 	"context"
 
 	"go.anx.io/go-anxcloud/pkg/client"
+	"go.anx.io/go-anxcloud/pkg/genericresource"
 	"go.anx.io/go-anxcloud/pkg/utils/test"
 
 	lbaasBackend "go.anx.io/go-anxcloud/pkg/lbaas/backend"
@@ -20,7 +21,7 @@ const loadbalancerIdentifier = "fc5d7390e9e4400a9efc73b4d8e0613a"
 
 var _ = Describe("lbaas/server client", Label("old client", "slow"), func() {
 	var cli client.Client
-	var api API
+	var api genericresource.API[Server, Definition]
 	var backend lbaasBackend.Backend
 
 	BeforeEach(func() {
